+++ optisim/mpc/__init__.py
@@ -1,0 +1,21 @@
+"""Model predictive control public exports."""
+
+from optisim.mpc.controller import (
+    FootstepPlan,
+    FootstepPlanner,
+    HumanoidMPC,
+    LinearMPC,
+    MPCConfig,
+    MPCSolution,
+    build_humanoid_mpc,
+)
+
+__all__ = [
+    "MPCConfig",
+    "MPCSolution",
+    "LinearMPC",
+    "HumanoidMPC",
+    "FootstepPlan",
+    "FootstepPlanner",
+    "build_humanoid_mpc",
+]